+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## [0.3] - 2020-10-10
+### Changed
+- `LoggerFactory`: added `getNullLoggerFactory()`
+- `ConsoleLogger`: improved options
+- `README.md`: improved documentation
+### Added
+- `Slf4JLoggerException`: handle errors in accessing slf4j
+
 ## [0.2] - 2020-10-06
 ### Changed
 - `Logger` (and implementing classes): message is now an `Object`; the framework will apply `toString()`